+++ src/ej_tools/routes.py
@@ -1,11 +1,10 @@
 import json
-import requests
 from boogie.router import Router
 from django.utils.translation import ugettext_lazy as _
 from django.http import HttpResponse
 from django.shortcuts import redirect
 from django.core.exceptions import ValidationError
-from .utils import npm_version, user_can_add_new_domain, prepare_host_with_https, get_host_with_protocol
+from .utils import npm_version, user_can_add_new_domain, prepare_host_with_https, get_host_with_schema
 from .forms import (
     RasaConversationForm,
     ConversationComponentForm,
@@ -13,25 +12,22 @@ from .forms import (
     MailingToolForm,
     MauticConversationForm,
 )
-from ej_conversations.models import Conversation
-from .models import (
+from ej_tools.models import (
     RasaConversation,
     ConversationMautic,
     MauticOauth2Service,
     MauticClient,
-    ChatbotTelegramTool,
-    ChatbotWhatsappTool,
     WebchatHelper,
 )
 from ej_conversations import models
-from .table import Tools
+from ej_signatures.models import SignatureFactory
 
 
 app_name = "ej_tools"
 urlpatterns = Router(
     template="ej_tools/{name}.jinja2",
     models={
-        "conversation": Conversation,
+        "conversation": models.Conversation,
         "connection": RasaConversation,
         "mautic_connection": ConversationMautic,
     },
@@ -42,13 +38,18 @@ conversation_tools_chatbot_url = f"{conversation_tools_url}/chatbot"
 
 @urlpatterns.route(conversation_tools_url, perms=["ej.can_edit_conversation:conversation"])
 def index(request, board, conversation, slug, npm=npm_version):
-    tools = Tools(conversation)
-    return {"tools": tools.list(), "conversation": conversation}
+    user_signature = SignatureFactory.get_user_signature(conversation.author)
+    tools = user_signature.get_conversation_tools(conversation)
+    return {"tools": tools, "conversation": conversation}
 
 
 @urlpatterns.route(conversation_tools_url + "/mailing")
 def mailing(request, board, conversation, slug):
     from .mailing import TemplateGenerator
+
+    user_signature = SignatureFactory.get_user_signature(conversation.author)
+    tool = user_signature.get_tool(_("Mailing campaign"), conversation)
+    tool.raise_error_if_not_active()
 
     template = "null"
     form = MailingToolForm(request.POST, conversation_id=conversation)
@@ -62,10 +63,9 @@ def mailing(request, board, conversation, slug):
             return response
         if "preview" in request.POST:
             template = json.dumps(template, ensure_ascii=False)
-    tools = Tools(conversation)
     return {
         "conversation": conversation,
-        "tool": tools.get(_("Mailing campaign")),
+        "tool": tool,
         "template_preview": template,
         "form": form,
     }
@@ -73,22 +73,21 @@ def mailing(request, board, conversation, slug):
 
 @urlpatterns.route(conversation_tools_url + "/opinion-component")
 def opinion_component(request, conversation, **kwargs):
-    from django.conf import settings
+    user_signature = SignatureFactory.get_user_signature(conversation.author)
+    tool = user_signature.get_tool(_("Opinion component"), conversation)
+    tool.raise_error_if_not_active()
 
-    schema = "https" if settings.ENVIRONMENT != "local" else "http"
     form = ConversationComponentForm(request.POST)
     conversation_component = ConversationComponent(form)
-    tools = Tools(conversation)
-
     if "preview" in request.POST:
         form.is_valid()
-        request.session["theme"] = form.cleaned_data["theme"] or "icd"
-        request.session["authentication_type"] = form.cleaned_data["authentication_type"]
-        return redirect(conversation.url("conversation-tools:opinion-component-preview"))
-
+        theme = form.cleaned_data["theme"]
+        if theme:
+            theme = f"?theme={theme}"
+        return redirect(conversation.url("conversation-tools:opinion-component-preview") + theme)
     return {
-        "schema": schema,
-        "tool": tools.get(_("Opinion component")),
+        "ej_domain": get_host_with_schema(request),
+        "tool": tool,
         "npm_version": npm_version(),
         "conversation": conversation,
         "form": form,
@@ -97,55 +96,50 @@ def opinion_component(request, conversation, **kwargs):
 
 
 @urlpatterns.route(conversation_tools_url + "/opinion-component/preview")
-def opinion_component_preview(request, board, conversation, slug):
-    host = get_host_with_protocol(request)
-    theme = request.session.get("theme")
-    auth_type = request.session.get("authentication_type")
+def opinion_component_preview(request, conversation, **kwargs):
+    user_signature = SignatureFactory.get_user_signature(conversation.author)
+    tool = user_signature.get_tool(_("Opinion component"), conversation)
+    tool.raise_error_if_not_active()
+
+    host = get_host_with_schema(request)
+    theme = request.GET.get("theme", "icd")
     return {
         "conversation": conversation,
-        "authentication_type": auth_type,
         "theme": theme,
         "host": host,
     }
 
 
 @urlpatterns.route(conversation_tools_url + "/chatbot")
-def chatbot(request, board, conversation, slug):
-    tools = Tools(conversation)
-    return {
-        "conversation": conversation,
-        "tool": tools.get(_("Chatbot")),
-    }
+def chatbot(request, conversation, **kwargs):
+    user_signature = SignatureFactory.get_user_signature(conversation.author)
+    return {"conversation": conversation, "tool": user_signature.get_tool(_("Opinion Bots"), conversation)}
 
 
 @urlpatterns.route(conversation_tools_url + "/telegram")
-def telegram(request, board, conversation, slug):
-    tools = Tools(conversation)
-
-    return {
-        "conversation": conversation,
-        "tool": tools.get(_("Chatbot")),
-        "channels": ChatbotTelegramTool.CHANNELS_CHOICES,
-        "shareText": ChatbotTelegramTool.SHARE,
-    }
+def telegram(request, conversation, **kwargs):
+    user_signature = SignatureFactory.get_user_signature(conversation.author)
+    tool = user_signature.get_tool(_("Opinion Bots"), conversation).telegram
+    tool.raise_error_if_not_active()
+    return {"conversation": conversation, "tool": tool}
 
 
 @urlpatterns.route(conversation_tools_chatbot_url + "/whatsapp")
 def whatsapp(request, board, conversation, slug):
-    tools = Tools(conversation)
-    return {
-        "conversation": conversation,
-        "tool": tools.get(_("Chatbot")),
-        "channels": ChatbotWhatsappTool.CHANNEL_CHOICES,
-        "shareText": ChatbotWhatsappTool.SHARE,
-    }
+    user_signature = SignatureFactory.get_user_signature(conversation.author)
+    tool = user_signature.get_tool(_("Opinion Bots"), conversation).whatsapp
+    tool.raise_error_if_not_active()
+    return {"conversation": conversation, "tool": tool}
 
 
 @urlpatterns.route(conversation_tools_chatbot_url + "/webchat")
 def webchat(request, conversation, **kwargs):
+    user_signature = SignatureFactory.get_user_signature(conversation.author)
+    tool = user_signature.get_tool(_("Opinion Bots"), conversation).webchat
+    tool.raise_error_if_not_active()
+
     user_can_add = user_can_add_new_domain(request.user, conversation)
-    host = get_host_with_protocol(request)
-    request.session["host"] = host
+    host = get_host_with_schema(request)
     webchat_preview_url = host + conversation.url("conversation-tools:webchat-preview")
 
     if "webchat-preview" in request.POST:
@@ -163,11 +157,10 @@ def webchat(request, conversation, **kwargs):
         form = RasaConversationForm()
 
     conversation_rasa_connections = models.RasaConversation.objects.filter(conversation=conversation)
-    tools = Tools(conversation)
     return {
         "conversation": conversation,
         "conversation_rasa_connections": conversation_rasa_connections,
-        "tool": tools.get(_("Chatbot")),
+        "tool": tool,
         "form": form,
         "is_valid_user": user_can_add,
     }
@@ -175,7 +168,11 @@ def webchat(request, conversation, **kwargs):
 
 @urlpatterns.route(conversation_tools_chatbot_url + "/webchat/preview")
 def webchat_preview(request, board, conversation, slug):
-    host = request.session.get("host")
+    user_signature = SignatureFactory.get_user_signature(conversation.author)
+    tool = user_signature.get_tool(_("Opinion Bots"), conversation).webchat
+    tool.raise_error_if_not_active()
+
+    host = get_host_with_schema(request)
     rasa_domain = WebchatHelper.get_rasa_domain(host)
     return {"conversation": conversation, "rasa_domain": rasa_domain}
 
@@ -183,6 +180,9 @@ def webchat_preview(request, board, conversation, slug):
 @urlpatterns.route(conversation_tools_chatbot_url + "/webchat/delete/<model:connection>")
 def delete_connection(request, board, conversation, slug, connection):
     user = request.user
+    user_signature = SignatureFactory.get_user_signature(conversation.author)
+    tool = user_signature.get_tool(_("Opinion Bots"), conversation).webchat
+    tool.raise_error_if_not_active()
 
     if user.is_staff or user.is_superuser or connection.conversation.author.id == user.id:
         connection.delete()
@@ -195,9 +195,13 @@ def delete_connection(request, board, conversation, slug, connection):
 
 
 @urlpatterns.route(
-    conversation_tools_chatbot_url + "/mautic", perms=["ej.can_access_mautic_connection:conversation"]
+    conversation_tools_url + "/mautic", perms=["ej.can_access_mautic_connection:conversation"]
 )
 def mautic(request, board, conversation, slug, oauth2_code=None):
+    user_signature = SignatureFactory.get_user_signature(conversation.author)
+    tool = user_signature.get_tool(_("Mautic"), conversation)
+    tool.raise_error_if_not_active()
+
     error_message = None
     connections = None
 
@@ -209,7 +213,6 @@ def mautic(request, board, conversation, slug, oauth2_code=None):
     except Exception as e:
         print(e)
 
-    tools = Tools(conversation)
     conversation_kwargs = {
         "conversation": conversation,
     }
@@ -236,7 +239,7 @@ def mautic(request, board, conversation, slug, oauth2_code=None):
     return {
         "conversation": conversation,
         "connections": connections,
-        "tool": tools.get(_("Mautic")),
+        "tool": tool,
         "form": form,
         "errors": error_message,
     }
@@ -247,6 +250,10 @@ def mautic(request, board, conversation, slug, oauth2_code=None):
     perms=["ej.can_access_mautic_connection:conversation"],
 )
 def delete_mautic_connection(request, board, conversation, slug, mautic_connection):
+    user_signature = SignatureFactory.get_user_signature(conversation.author)
+    tool = user_signature.get_tool(_("Mautic"), conversation)
+    tool.raise_error_if_not_active()
+
     mautic_connection.delete()
     return redirect(conversation.url("conversation-tools:mautic"))